+++ spec/cuprum/rails/actions/destroy_spec.rb
@@ -9,9 +9,7 @@ require 'support/tome'
 RSpec.describe Cuprum::Rails::Actions::Destroy do
   include Cuprum::Rails::RSpec::Actions::DestroyContracts
 
-  subject(:action) do
-    described_class.new(repository: repository, resource: resource)
-  end
+  subject(:action) { described_class.new }
 
   let(:repository) { Cuprum::Rails::Repository.new }
   let(:resource) do