+++ spec/cuprum/rails/actions/new_spec.rb
@@ -8,9 +8,7 @@ require 'support/book'
 RSpec.describe Cuprum::Rails::Actions::New do
   include Cuprum::Rails::RSpec::Actions::NewContracts
 
-  subject(:action) do
-    described_class.new(repository: repository, resource: resource)
-  end
+  subject(:action) { described_class.new }
 
   let(:repository) { Cuprum::Rails::Repository.new }
   let(:resource) do