+++ spec/cuprum/rails/controllers/class_methods/actions_spec.rb
@@ -2,6 +2,7 @@
 
 require 'cuprum/rails/controllers/class_methods/actions'
 require 'cuprum/rails/controllers/class_methods/configuration'
+require 'cuprum/rails/controllers/class_methods/middleware'
 require 'cuprum/rails/controllers/class_methods/validations'
 
 require 'support/examples/controller_examples'
@@ -18,7 +19,15 @@ RSpec.describe Cuprum::Rails::Controllers::ClassMethods::Actions do
   example_class 'Spec::Controller', Struct.new(:request) do |klass|
     klass.extend Cuprum::Rails::Controllers::ClassMethods::Actions # rubocop:disable RSpec/DescribedClass
     klass.extend Cuprum::Rails::Controllers::ClassMethods::Configuration
+    klass.extend Cuprum::Rails::Controllers::ClassMethods::Middleware
     klass.extend Cuprum::Rails::Controllers::ClassMethods::Validations
+
+    klass.define_method(:action_options) do
+      {
+        repository: Spec::Controller.repository,
+        resource:   Spec::Controller.resource
+      }
+    end
 
     klass.define_method(:controller_name) { 'api/books' }
   end