+++ spec/support/actions/destroy_book.rb
@@ -51,7 +51,7 @@ module Spec::Support::Actions
       end
     end
 
-    def process(request:)
+    def process(**)
       @chapters = nil
 
       super