+++ spec/support/actions/login.rb
@@ -18,7 +18,7 @@ module Spec::Support::Actions
 
     private
 
-    def process(request:)
+    def process(**)
       super
 
       { 'ok' => true }